[favourites] Refactor GUI actions.
