+++ xbmc/favourites/CMakeLists.txt
@@ -3,13 +3,15 @@ set(SOURCES ContextMenus.cpp
             GUIViewStateFavourites.cpp
             GUIWindowFavourites.cpp
             FavouritesService.cpp
-            FavouritesURL.cpp)
+            FavouritesURL.cpp
+            FavouritesUtils.cpp)
 
 set(HEADERS ContextMenus.h
             GUIDialogFavourites.h
             GUIViewStateFavourites.h
             GUIWindowFavourites.h
             FavouritesService.h
-            FavouritesURL.h)
+            FavouritesURL.h
+            FavouritesUtils.h)
 
 core_add_library(favourites)
