+++ xbmc/favourites/GUIWindowFavourites.cpp
@@ -10,19 +10,16 @@
 
 #include "FileItem.h"
 #include "ServiceBroker.h"
-#include "dialogs/GUIDialogFileBrowser.h"
 #include "favourites/FavouritesURL.h"
+#include "favourites/FavouritesUtils.h"
 #include "guilib/GUIComponent.h"
-#include "guilib/GUIKeyboardFactory.h"
+#include "guilib/GUIMessage.h"
 #include "guilib/GUIWindowManager.h"
-#include "guilib/LocalizeStrings.h"
 #include "input/actions/Action.h"
 #include "input/actions/ActionIDs.h"
 #include "messaging/ApplicationMessenger.h"
-#include "storage/MediaManager.h"
 #include "utils/PlayerUtils.h"
 #include "utils/StringUtils.h"
-#include "view/GUIViewState.h"
 
 CGUIWindowFavourites::CGUIWindowFavourites()
   : CGUIMediaWindow(WINDOW_FAVOURITES, "MyFavourites.xml")
@@ -103,12 +100,12 @@ bool CGUIWindowFavourites::OnAction(const CAction& action)
   }
   else if (action.GetID() == ACTION_MOVE_ITEM_UP)
   {
-    if (ShouldEnableMoveItems())
+    if (FAVOURITES_UTILS::ShouldEnableMoveItems())
       return MoveItem(selectedItem, -1);
   }
   else if (action.GetID() == ACTION_MOVE_ITEM_DOWN)
   {
-    if (ShouldEnableMoveItems())
+    if (FAVOURITES_UTILS::ShouldEnableMoveItems())
       return MoveItem(selectedItem, +1);
   }
   else if (action.GetID() == ACTION_DELETE_ITEM)
@@ -158,52 +155,12 @@ bool CGUIWindowFavourites::Update(const std::string& strDirectory,
   return CGUIMediaWindow::Update(directory, updateFilterPath);
 }
 
-bool CGUIWindowFavourites::ChooseAndSetNewName(CFileItem& item)
-{
-  std::string label = item.GetLabel();
-  if (CGUIKeyboardFactory::ShowAndGetInput(label, CVariant{g_localizeStrings.Get(16008)},
-                                           false)) // Enter new title
-  {
-    item.SetLabel(label);
-    return true;
-  }
-  return false;
-}
-
-bool CGUIWindowFavourites::ChooseAndSetNewThumbnail(CFileItem& item)
-{
-  CFileItemList prefilledItems;
-  if (item.HasArt("thumb"))
-  {
-    const auto current = std::make_shared<CFileItem>("thumb://Current", false);
-    current->SetArt("thumb", item.GetArt("thumb"));
-    current->SetLabel(g_localizeStrings.Get(20016)); // Current thumb
-    prefilledItems.Add(current);
-  }
-
-  const auto none = std::make_shared<CFileItem>("thumb://None", false);
-  none->SetArt("icon", item.GetArt("icon"));
-  none->SetLabel(g_localizeStrings.Get(20018)); // No thumb
-  prefilledItems.Add(none);
-
-  std::string thumb;
-  VECSOURCES sources;
-  CServiceBroker::GetMediaManager().GetLocalDrives(sources);
-  if (CGUIDialogFileBrowser::ShowAndGetImage(prefilledItems, sources, g_localizeStrings.Get(1030),
-                                             thumb)) // Browse for image
-  {
-    item.SetArt("thumb", thumb);
-    return true;
-  }
-  return false;
-}
-
 bool CGUIWindowFavourites::MoveItem(int item, int amount)
 {
   if (item < 0 || item >= m_vecItems->Size() || m_vecItems->Size() < 2 || amount == 0)
     return false;
 
-  if (MoveItem(*m_vecItems, (*m_vecItems)[item], amount) &&
+  if (FAVOURITES_UTILS::MoveItem(*m_vecItems, (*m_vecItems)[item], amount) &&
       CServiceBroker::GetFavouritesService().Save(*m_vecItems))
   {
     int selected = item + amount;
@@ -219,73 +176,14 @@ bool CGUIWindowFavourites::MoveItem(int item, int amount)
   return false;
 }
 
-bool CGUIWindowFavourites::MoveItem(CFileItemList& items,
-                                    const std::shared_ptr<CFileItem>& item,
-                                    int amount)
-{
-  if (items.Size() < 2 || amount == 0)
-    return false;
-
-  int itemPos = -1;
-  for (const auto& i : items)
-  {
-    itemPos++;
-
-    if (i->GetPath() == item->GetPath())
-      break;
-  }
-
-  if (itemPos < 0 || itemPos >= items.Size())
-    return false;
-
-  int nextItem = (itemPos + amount) % items.Size();
-  if (nextItem < 0)
-  {
-    const auto itemToAdd(item);
-    items.Remove(itemPos);
-    items.Add(itemToAdd);
-  }
-  else if (nextItem == 0)
-  {
-    const auto itemToAdd(item);
-    items.Remove(itemPos);
-    items.AddFront(itemToAdd, 0);
-  }
-  else
-  {
-    items.Swap(itemPos, nextItem);
-  }
-  return true;
-}
-
 bool CGUIWindowFavourites::RemoveItem(int item)
 {
   if (item < 0 || item >= m_vecItems->Size())
     return false;
 
-  if (RemoveItem(*m_vecItems, (*m_vecItems)[item]) &&
+  if (FAVOURITES_UTILS::RemoveItem(*m_vecItems, (*m_vecItems)[item]) &&
       CServiceBroker::GetFavouritesService().Save(*m_vecItems))
     return true;
 
   return false;
-}
-
-bool CGUIWindowFavourites::RemoveItem(CFileItemList& items, const std::shared_ptr<CFileItem>& item)
-{
-  int iBefore = items.Size();
-  items.Remove(item.get());
-  return items.Size() == iBefore - 1;
-}
-
-bool CGUIWindowFavourites::ShouldEnableMoveItems()
-{
-  auto& mgr = CServiceBroker::GetGUI()->GetWindowManager();
-  CGUIWindowFavourites* window = mgr.GetWindow<CGUIWindowFavourites>(WINDOW_FAVOURITES);
-  if (window && window->IsActive())
-  {
-    const CGUIViewState* state = window->GetViewState();
-    if (state && state->GetSortMethod().sortBy != SortByUserPreference)
-      return false; // in favs window, allow move only if current sort method is by user preference
-  }
-  return true;
 }
